requirements.txt
pandas==2.2.2
numpy==1.26.4
ijson==3.2.3
matplotlib==3.8.4
python-dateutil==2.9.0.post0
reportlab==4.2.2
